requirements.txt
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
gunicorn==21.2.0
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
langdetect==1.0.9
MarkupSafe==2.1.5
matplotlib==3.8.3
mplcursors==0.5.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyclipper==1.3.0.post5
pycparser==2.21
pydyf==0.9.0
pyparsing==3.1.2
pypdfium2==4.27.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-doctr==0.8.1
PyYAML==6.0.1
rapidfuzz==3.6.2
requests==2.31.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==2.2.1
weasyprint==61.1
webencodings==0.5.1
Werkzeug==3.0.1
zipp==3.17.0
zopfli==0.2.3
